Add files via upload
This repository contains my extensive System Analysis and Design for a Medium Enterprise named Thompson Cycles. It demonstrates Software Development Methodology selection for the project, Requirement Elicitation, Detailed Class and Entity Relationship Diagrams with the Database Normalisation Tables. Also included is the User Interface Mock-ups and a detailed explanation of the components. Lastly, I designed a detailed 3-D and 2-D Deployment Diagram which offers a very impressive view of the System Infrastructure and architecture.